move concern to assignment stats
