+++ app/models/assignment.rb
@@ -1,7 +1,7 @@
 class Assignment < ActiveRecord::Base
   include Copyable
   include Gradable
-  include Analysable
+  include AssignmentStats
   include MultipleFileAttributes
   include Sanitizable
   include ScoreLevelable
@@ -285,11 +285,6 @@ class Assignment < ActiveRecord::Base
 
   def grade_import(students, options = {})
     GradeExporter.new.export_grades(self, students, options)
-  end
-
-  # Creating an array with the set of pass/fail statuses earned on the assignment
-  def percentage_pass_fail_earned
-    { scores: earned_status_count.collect { |s| { data: s[1], name: s[0] }}}
   end
 
   private
